+++ pkg/client/common.go
@@ -1,0 +1,6 @@
+package client
+
+type contextKey string
+
+// TraceIDContextKey is the key to store the trace ID in the context
+const TraceIDContextKey contextKey = "trace-id"