+++ CHANGELOG.md
@@ -9,9 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Client for `user-preferences` service
+- Logging functions accepting context e.g., `LogInfofCtx`, `LogErrorfCtx`
+
 ### Changed
 
+- Update client for `cds-notification` service `>= v0.6.x`
+
 ### Deprecated
+
+- Deprecate client interfaces `NotificationV1` and `NotificationV2`.
+    Please migrate to compatible `NotificationV3` client.
+    When using this version of `go-svc`, all `NotificationClient`s that rely on `cds-notification` `< v0.6.0` should be changed to `NotificationClientLegacy`
+    or updated migrated to support `NotificationV3` (it adds one optional parameter).
+- Logging functions without context, e.g., `LogInfo`, `LogError`, `LogWarning`
 
 ### Removed